requirements.txt
astroid==2.4.2
attrs==19.3.0
aws-cdk.assets==1.55.0
aws-cdk.aws-cloudwatch==1.55.0
aws-cdk.aws-codeguruprofiler==1.55.0
aws-cdk.aws-ec2==1.55.0
aws-cdk.aws-efs==1.55.0
aws-cdk.aws-events==1.55.0
aws-cdk.aws-iam==1.55.0
aws-cdk.aws-kms==1.55.0
aws-cdk.aws-lambda==1.55.0
aws-cdk.aws-logs==1.55.0
aws-cdk.aws-s3==1.55.0
aws-cdk.aws-s3-assets==1.55.0
aws-cdk.aws-sns==1.55.0
aws-cdk.aws-sns-subscriptions==1.55.0
aws-cdk.aws-sqs==1.55.0
aws-cdk.aws-ssm==1.55.0
aws-cdk.cloud-assembly-schema==1.55.0
aws-cdk.core==1.55.0
aws-cdk.cx-api==1.55.0
aws-cdk.region-info==1.55.0
awswrangler==1.6.3
backcall==0.2.0
boto3==1.14.22
botocore==1.17.22
cattrs==1.0.0
cdk-constants==3.0.3
certifi==2020.6.20
chardet==3.0.4
constructs==3.0.4
decorator==4.4.2
docutils==0.15.2
fsspec==0.7.4
idna==2.10
ipython==7.16.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.2
jmespath==0.10.0
jsii==1.9.0
lazy-object-proxy==1.4.3
loguru==0.5.1
mccabe==0.6.1
more-itertools==8.4.0
numpy==1.19.0
packaging==20.4
pandas==1.0.5
pandas-market-calendars==1.4.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
publication==0.0.3
py==1.9.0
pyarrow==0.17.1
Pygments==2.6.1
pylint==2.5.3
PyMySQL==0.9.3
pyparsing==2.4.7
pypng==0.0.20
pytest==5.4.3
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
s3fs==0.4.2
s3transfer==0.3.3
six==1.15.0
snappy-manifolds==1.1.1
spherogram==1.8.3
SQLAlchemy==1.3.13
sqlalchemy-redshift==0.8.1
-e git+git@github.com:Haklim733/stocks.git@f20e73f43eff1eeef459d9c520a5eadbef4f1b16#egg=stocks_cdk
toml==0.10.1
traitlets==4.3.3
typing-extensions==3.7.4.2
urllib3==1.25.9
wcwidth==0.2.5
wrapt==1.12.1
yapf==0.30.0
